stats now working, with minor bug

+++ par-shell-terminal.c
@@ -27,10 +27,10 @@ Main thread
 */
 int main(int argc, char* argv[]){
 
-	char *my_string, *message, *pipeout;
+	char *my_string, *message, pipeout[20];
 	my_string = (char*) malloc(NARGUMENTOS*sizeof(char));
 	message = (char*) malloc(DIM*sizeof(char));
-	pipeout = (char*) malloc(DIM*sizeof(char));
+	//pipeout = (char*) malloc(DIM*sizeof(char));
 	int fserv, fclient;
 	size_t nbytes = DIM;
 	sprintf(pipeout, "par-shell-%d", getpid());
@@ -53,11 +53,16 @@ int main(int argc, char* argv[]){
 
 		if (strcmp(my_string,"stats\n")==0) {
 			printf("entrei no stats\n");
-			strtok(my_string, "\n");
-			strcat(my_string, " ");
-			strcat(my_string,pipeout);
+			//strtok(my_string, "\n");
+			//strcat(my_string, " ");
+			//strcat(my_string,pipeout);
+			//strcat(my_string, "\n");
 			printf("%s\n",my_string );
 
+			if((sprintf(my_string, "stats %s\n", pipeout)) <0){
+				perror("Error at sprintf");
+				exit(EXIT_FAILURE);
+			}
 			unlink(pipeout);
 
 			if (mkfifo (pipeout, PREMISSOES) < 0) {
@@ -70,7 +75,8 @@ int main(int argc, char* argv[]){
 				exit(EXIT_FAILURE);
 			}
 
-				printf("%s\n",my_string );
+			printf("%s\n",my_string );
+			printf("pipeout:%s\n",pipeout );
 			if((fclient = open(pipeout, O_RDONLY)) <0 ){
 				perror("Error associating FIFO in par-shell");
 				exit(EXIT_FAILURE);
@@ -79,12 +85,13 @@ int main(int argc, char* argv[]){
 			printf("escrevi para a par-shell\n");
 			printf("abri e vou ler\n");
 			printf("%s-1\n",message );
-			
-			if( (read(fclient,message,strlen(message))) < 0){
+
+			if( read(fclient,message,200) < 0){
 				perror("Error writing stream");
 				exit(EXIT_FAILURE);
 			}
-			printf("%s-2\n",message);
+			printf("message:%s\n",message);
+			unlink(pipeout);
 		}
 		if (strcmp(my_string,"exit\n")==0) {
 			unlink(pipeout);
